separated spmat.go from base.go
